Move FunctionsStringHashFixedString template to single file
